requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

albumentations==2.0.8
kaggle==1.7.4.5
kagglehub==1.0.0
matplotlib==3.10.9
numpy==2.2.6
opencv-python==4.13.0.92
pandas==2.3.3
pytest==9.0.3
PyYAML==6.0.3
scikit-learn==1.7.2
swanlab==0.7.16
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.9.0
torchvision==0.20.1+cu121
tqdm==4.67.3
ultralytics==8.4.41
wandb==0.26.1
